adb: fix 64-bit build
strlen returns a size_t, but the * modifier in printf expects an int.
On arm64 size_t != int.

(cherry-picked from commit 6d6a898b0c4e0283569209ba74bd7dbf49da73a2)

+++ adb/commandline.c
@@ -601,7 +601,7 @@ int adb_sideload_host(const char* fn) {
         }
     }
 
-    printf("\rTotal xfer: %.2fx%*s\n", (double)xfer / (sz ? sz : 1), strlen(fn)+10, "");
+    printf("\rTotal xfer: %.2fx%*s\n", (double)xfer / (sz ? sz : 1), (int)strlen(fn)+10, "");
 
   done:
     if (fd >= 0) adb_close(fd);
